refactor(api): move in memory temporary storage to src/shared
